Rewrites the main so that it uses Windows API
The WinAPI will be needed for future uses of the project, so there needs
to be some way of starting the project. The `add_compile_options` is
taken straight from CMake's documentation - it sets flags for Unix
systems, but as of now, I do not plan to create this application for
Unix systems. I will try to do it after I'm done with at least one
platform :)