Moves custom flag settings to the private build script instead of requiring them on project level
From what I've learned today, it's ill advised to use `add_compile_flags` for things like setting warning level or enabling debug information. It's cleaner to set them in the building script and keep it as a local configuration. `ninja` invocation has been changed to use `cmake`, as I can manipulate the configuration that way and it also highlights the basic use of CMake - first `cmake` configures the project and the second one builds it.